requirements.txt
behave
nose
pyserial
celery

# sudo apt-get install arduino
